# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a todo-list server.",
        "title": "Gin Swagger todo-list API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/addItem": {
            "post": {
                "description": "add task to given user",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "addItem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Item",
                        "name": "Item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.itemRequirements"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/addUser": {
            "post": {
                "description": "add user with given username and password",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "addUser",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "Item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.userRequirements"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/getItemById/{id}": {
            "get": {
                "description": "get tasks of given user ID and task ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "getItemById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/getItems": {
            "get": {
                "description": "get tasks of given user ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "getItems",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter",
                        "name": "rawName",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/getItemsByCount": {
            "get": {
                "description": "get a certain number of tasks of given user ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "getItemsByCount",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter",
                        "name": "rawName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Count",
                        "name": "count",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/removeItem/{id}": {
            "delete": {
                "description": "Remove tasks with given ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "removeItem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/updateItem/{id}": {
            "patch": {
                "description": "Update tasks with given ID",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "patchItem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Name",
                        "name": "Item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.itemRequirements"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/validateUser": {
            "post": {
                "description": "validate user with given username and passwod",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "validateUser",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "Item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.userRequirements"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.itemRequirements": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "topPriority": {
                    "type": "boolean"
                }
            }
        },
        "main.userRequirements": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}